scripts/species/data-prep/requirements.txt
pygbif
qwikidata
pandas
tqdm
jupyterlab
openpyxl
google-cloud-storage